+++ tests/calculator/controller/key_char_tilde.cpp
@@ -35,6 +35,7 @@ TEST(controller, key_char_tilde_too_few_elements) {
 
 TEST(controller, key_char_tilde_input) {
   tmodel model;
+  model.grouping_toggle();
   model.base_set(tbase::binary);
   tcontroller controller{model};
   model.input_append("3");
@@ -49,6 +50,7 @@ TEST(controller, key_char_tilde_input) {
 
 TEST(controller, key_char_tilde_stack) {
   tmodel model;
+  model.grouping_toggle();
   model.base_set(tbase::binary);
   tcontroller controller{model};
   handle_input(controller, model, "3");
@@ -63,6 +65,7 @@ TEST(controller, key_char_tilde_stack) {
 
 TEST(controller, key_char_tilde_diagnostics_cleared) {
   tmodel model;
+  model.grouping_toggle();
   model.base_set(tbase::binary);
   tcontroller controller{model};
   model.diagnostics_set("Cleared");
@@ -82,7 +85,8 @@ TEST(controller, key_char_tilde_input_invalid) {
   model.input_append("abc");
 
   controller.handle_keyboard_input(tmodifiers::none, '~');
-  EXPECT_EQ(model.diagnostics_get(), "Invalid numeric value or command");
+  EXPECT_EQ(model.diagnostics_get(),
+            format_error("Invalid numeric value or command"));
   EXPECT_TRUE(model.stack().empty());
   EXPECT_EQ(model.input_get(), "abc");
 }